split main config into multiple files
